requirements.txt
altgraph==0.17.4
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.1
future==1.0.0
macholib==1.16.3
pillow==10.3.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
